Update hypothesis.Rmd
A small addition to the hypothesis section. Not comprehensive, please review carefully before merge (it might needs rewriting). But I believe for students it's important to know that (1) stats test are there to evaluate hypotheses; (2) usually we contrast two hypotheses with each other. I intentionally left out the whole frequentist/Bayesian discussion and that we reject H0s, that is something people could add in the NHST section or a Bayesian section if it is added at some point.